Actually write to the sqe index array, as intended.
This bug went unnoticed for a while because it wrote to `sq_entries`,
which would be immediately overwritten in normal execution. This bug
caused only the 0th entry to be submitted, and to be submitted multiple
times, once for entry that *should* be submitted. Thank you "Lord of the
io_uring" for giving me a reference to diff against!