Split out new `Op` type to handle generic polling/ring logic.
Also, switch to a monotonic increasing ticket for `user_data` rather than
pointers since we can drop futures and reuse addresses leading to reused
`user_data`. Reuse break the assumption that `sqe.user_data ==
cqe.user_data` implies the cqe is related to the sqe, which could cause
safety issues if I'm not programming defensively for the rest of the
ops.