Fix unsoundness on drop of ops.
`drop(x)` ends the lifetime of x, but in our case `Uring` and
potentially even the kernel still hold a reference bound by the lifetime
of that object, leading to an easy use-after-free.

The solution is to submit a cancellation request on drop and block until
it completes. Because we might not be able to cancel a request even
while it's still not done, we also have to await the original op, though
that should be instant if it *was* canceled.

Problem example:
```rs
let mut buf = [0];
let read = Read::new(&ring, file, 0, buf).build();
drop(read);
// Exclusive reference of `buf` ends
ring.borrow_mut().submit();
// Oh no, the kernel just wrote to `buf`!
```

This commit also adds a unit test that tests basically the above
scenario.